Move localisation group in to its own class
- Also change mph option to `CheckboxOption`
