+++ TLM/TLM/State/OptionsTabs/GeneralTab_LocalisationGroup.cs
@@ -1,0 +1,194 @@
+namespace TrafficManager.State {
+    using ColossalFramework.UI;
+    using CSUtil.Commons;
+    using ICities;
+    using System.Collections.Generic;
+    using System.Linq;
+    using TrafficManager.Lifecycle;
+    using TrafficManager.State.ConfigData;
+    using TrafficManager.UI;
+    using TrafficManager.UI.Helpers;
+    using TrafficManager.UI.Textures;
+
+    /// <summary>
+    /// Adds localisation options to General options tab.
+    /// </summary>
+    public static class GeneralTab_LocalisationGroup {
+
+        // TODO: Implement global config updates direct from CheckboxOption
+        public static CheckboxOption DisplaySpeedLimitsMph =
+            new ("DebugCheckboxA", Options.PersistTo.None) {
+                Label = "Checkbox:Display speed limits mph",
+                Translator = TSpeedLimits,
+                Handler = OnDisplaySpeedLimitsMphChanged,
+            };
+
+        private static UIDropDown _roadSignsThemeDropdown;
+
+        internal static void AddUI(UIHelperBase tab) {
+
+            var group = tab.AddGroup(T("General.Group:Localisation"));
+
+            AddLanguageDropDown(group);
+
+            DisplaySpeedLimitsMph.Value = GlobalConfig.Instance.Main.DisplaySpeedLimitsMph;
+            DisplaySpeedLimitsMph.AddUI(group);
+
+            AddRoadSignThemeDropDown(group);
+        }
+
+        private static string T(string key) => Translation.Options.Get(key);
+
+        private static string TLang(string key, int langNum) =>
+            Translation.Options.Get(lang: Translation.AvailableLanguageCodes[langNum], key: key);
+
+        // TODO: Create a UI component helper for drop-down lists
+        private static void AddLanguageDropDown(UIHelperBase group) {
+            string[] languageLabels = new string[Translation.AvailableLanguageCodes.Count + 1];
+            languageLabels[0] = T("General.Dropdown.Option:Game language");
+
+            for (int i = 0; i < Translation.AvailableLanguageCodes.Count; ++i) {
+                languageLabels[i + 1] = TLang("General.Dropdown.Option:Language Name", i);
+            }
+
+            int languageIndex = 0;
+            string curLangCode = GlobalConfig.Instance.LanguageCode;
+
+            if (curLangCode != null) {
+                languageIndex = Translation.AvailableLanguageCodes.IndexOf(curLangCode);
+                if (languageIndex < 0) {
+                    languageIndex = 0;
+                } else {
+                    ++languageIndex;
+                }
+            }
+
+            group.AddDropdown(
+                text: T("General.Dropdown:Select language") + ":",
+                options: languageLabels,
+                defaultSelection: languageIndex,
+                eventCallback: OnLanguageChanged);
+        }
+
+        private static string TSpeedLimits(string key) => Translation.SpeedLimits.Get(key);
+
+        private static string TranslateThemeName(string themeName) => TSpeedLimits($"RoadSignTheme:{themeName}");
+
+        // TODO: Create a UI component helper for drop-down lists
+        private static void AddRoadSignThemeDropDown(UIHelperBase group) {
+            Main mainConfig = GlobalConfig.Instance.Main;
+
+            List<string> themeNames = RoadSignThemes.Instance.ThemeNames;
+            var themeOptions = themeNames.Select(TranslateThemeName).ToArray();
+            int selectedThemeIndex = themeNames.FindIndex(x => x == mainConfig.RoadSignTheme);
+            int defaultSignsThemeIndex = RoadSignThemes.Instance.FindDefaultThemeIndex(mainConfig.DisplaySpeedLimitsMph);
+
+            _roadSignsThemeDropdown = group.AddDropdown(
+                text: TSpeedLimits("General.Dropdown:Road signs theme") + ":",
+                options: themeOptions,
+                defaultSelection: selectedThemeIndex >= 0 ? selectedThemeIndex : defaultSignsThemeIndex,
+                eventCallback: OnRoadSignsThemeChanged) as UIDropDown;
+
+            _roadSignsThemeDropdown.width *= 2.0f;
+        }
+
+        private static void OnLanguageChanged(int newLanguageIndex) {
+            if (newLanguageIndex <= 0) {
+                // use game language
+                GlobalConfig.Instance.LanguageCode = null;
+                GlobalConfig.WriteConfig();
+
+                // TODO: Move this to the owner class and implement IObserver<ModUI.EventPublishers.LanguageChangeNotification>
+                Translation.SetCurrentLanguageToGameLanguage();
+                Options.RebuildMenu();
+            } else if (newLanguageIndex - 1 < Translation.AvailableLanguageCodes.Count) {
+                // use tmpe language
+                string newLang = Translation.AvailableLanguageCodes[newLanguageIndex - 1];
+                GlobalConfig.Instance.LanguageCode = newLang;
+                GlobalConfig.WriteConfig();
+
+                // TODO: Move this to the owner class and implement IObserver<ModUI.EventPublishers.LanguageChangeNotification>
+                Translation.SetCurrentLanguageToTMPELanguage();
+                Options.RebuildMenu();
+            } else {
+                Log.Warning($"GeneralTab.LocalisationGroup.onLanguageChanged({newLanguageIndex}): Invalid language index");
+                return;
+            }
+
+            if (TMPELifecycle.InGameOrEditor()) {
+                // Events will be null when mod is not fully loaded and language changed in main menu
+                ModUI.Instance.Events.LanguageChanged();
+            }
+
+            Options.RebuildOptions();
+        }
+
+        private static void OnDisplaySpeedLimitsMphChanged(bool newMphValue) {
+            bool supportedByTheme = newMphValue
+                                        ? RoadSignThemes.ActiveTheme.SupportsMph
+                                        : RoadSignThemes.ActiveTheme.SupportsKmph;
+
+            Main mainConfig = GlobalConfig.Instance.Main;
+
+            if (!supportedByTheme) {
+                // Reset to German road signs theme
+                _roadSignsThemeDropdown.selectedIndex = RoadSignThemes.Instance.FindDefaultThemeIndex(newMphValue);
+                mainConfig.RoadSignTheme = RoadSignThemes.Instance.GetDefaultThemeName(newMphValue);
+                Log.Info(
+                    $"Display MPH changed to {newMphValue}, but was not supported by current theme, "
+                    + "so theme was also reset to German_Kmph");
+            } else {
+                Log.Info($"Display MPH changed to {newMphValue}");
+            }
+
+            mainConfig.DisplaySpeedLimitsMph = newMphValue;
+            GlobalConfig.WriteConfig();
+
+            if (Options.IsGameLoaded(false)) {
+                ModUI.Instance.Events.DisplayMphChanged(newMphValue);
+            }
+        }
+
+        private static void OnRoadSignsThemeChanged(int newThemeIndex) {
+            if (!Options.IsGameLoaded()) {
+                return;
+            }
+
+            var newTheme = RoadSignThemes.Instance.ThemeNames[newThemeIndex];
+
+            Main mainConfig = GlobalConfig.Instance.Main;
+
+            switch (RoadSignThemes.Instance.ChangeTheme(
+                        newTheme: newTheme,
+                        mphEnabled: mainConfig.DisplaySpeedLimitsMph)) {
+                case RoadSignThemes.ChangeThemeResult.Success:
+                    Log.Info($"Road Sign theme changed to {newTheme}");
+                    mainConfig.RoadSignTheme = newTheme;
+                    break;
+                case RoadSignThemes.ChangeThemeResult.ForceKmph:
+                    mainConfig.DisplaySpeedLimitsMph = false;
+                    DisplaySpeedLimitsMph.Value = false;
+
+                    Log.Info($"Road Sign theme was changed to {newTheme} AND display switched to km/h");
+
+                    if (Options.IsGameLoaded(false)) {
+                        ModUI.Instance.Events.DisplayMphChanged(false);
+                    }
+                    break;
+                case RoadSignThemes.ChangeThemeResult.ForceMph:
+                    mainConfig.DisplaySpeedLimitsMph = true;
+                    DisplaySpeedLimitsMph.Value = true;
+
+                    Log.Info($"Road Sign theme was changed to {newTheme} AND display switched to MPH");
+
+                    if (Options.IsGameLoaded(false)) {
+                        ModUI.Instance.Events.DisplayMphChanged(true);
+                    }
+                    break;
+            }
+
+            mainConfig.RoadSignTheme = newTheme;
+            GlobalConfig.WriteConfig();
+        }
+    }
+}
